fix: resolve remaining format warnings across modules
Replace incorrect printf format specifiers that caused -Wformat warnings:
- %d -> %zu for size_t values
- %d -> static_cast<int>() for enum class values in log/assert messages
- %lld/%llu -> PRId64/PRIu64 macros (where not already covered)
- Remove extra format arguments
- Fix PRIXPTR usage in flv_private.h via reinterpret_cast<uintptr_t>
- Fix missing log arguments in physical_port.cpp

Result: 0 compile warnings across full clean build.

+++ src/projects/base/provider/pull_provider/application.cpp
@@ -156,14 +156,14 @@ namespace pvd
 
 						if((elapsed_time_from_last_sent > unused_stream_timeout_ms) && (!is_persistent))
 						{
-							logtw("%s/%s(%u) stream will be deleted because it hasn't been used for %u milliseconds", stream->GetApplicationInfo().GetVHostAppName().CStr(), stream->GetName().CStr(), stream->GetId(), elapsed_time_from_last_sent);
+							logtw("%s/%s(%u) stream will be deleted because it hasn't been used for %lld milliseconds", stream->GetApplicationInfo().GetVHostAppName().CStr(), stream->GetName().CStr(), stream->GetId(), static_cast<long long>(elapsed_time_from_last_sent));
 							DeleteStream(stream);
 						}
 						// The stream type is pull stream, if packets do NOT arrive for more than 3 seconds, it is a seriously warning situation
 						else if(elapsed_time_from_last_recv > no_input_timeout_ms && (!is_persistent))
 						{
-							logtw("Stop stream %s/%s(%u) : there are no incoming packets. %d milliseconds have elapsed since the last packet was received.",
-								  stream->GetApplicationInfo().GetVHostAppName().CStr(), stream->GetName().CStr(), stream->GetId(), elapsed_time_from_last_recv);
+							logtw("Stop stream %s/%s(%u) : there are no incoming packets. %lld milliseconds have elapsed since the last packet was received.",
+								  stream->GetApplicationInfo().GetVHostAppName().CStr(), stream->GetName().CStr(), stream->GetId(), static_cast<long long>(elapsed_time_from_last_recv));
 
 							// When the stream is stopped, it tries to reconnect using the next url.
 							stream->Stop();
